# schemas/generated/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Transflow loopback API",
    "version": "1"
  },
  "security": [
    {
      "bearer": []
    },
    {
      "session": []
    }
  ],
  "paths": {
...
    "/api/v1/capabilities": {
      "get": {
        "operationId": "get_api_v1_capabilities",
        "responses": {
          "200": {
            "description": "Context-bound result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "request_id",
                    "context",
                    "data"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "request_id": {
                      "$ref": "#/components/schemas/Uuid"
                    },
                    "context": {
                      "oneOf": [
                        {
                          "type": "null"
                        },
                        {
                          "$ref": "#/components/schemas/ApiContextV1"
                        }
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ApiCapabilitiesV1"
                    }
                  }
                },
                "example": {
                  "request_id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                  "context": null,
                  "data": {
                    "api_version": 1,
                    "protocol_major": 1,
                    "workspace": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                    "operations": [],
                    "engines": [],
                    "limits": {
                      "body_bytes": 0,
                      "in_flight": 0,
                      "page_default": 0,
                      "page_max": 0,
                      "graph_default": 0,
                      "graph_max": 0,
                      "graph_expansion_threshold": 0,
                      "source_bytes": 0
                    },
                    "metadata_reads_import_code": false,
                    "external_reads": "provider_owned_leased",
                    "schedules": false,
                    "ui": false
                  }
                }
              }
            }
          },
          "default": {
            "description": "Typed safe error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/api/v1/sessions/launch": {
      "post": {
        "operationId": "post_api_v1_sessions_launch",
        "responses": {
          "200": {
            "description": "Context-bound result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "request_id",
                    "context",
                    "data"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "request_id": {
                      "$ref": "#/components/schemas/Uuid"
                    },
                    "context": {
                      "oneOf": [
                        {
                          "type": "null"
                        },
                        {
                          "$ref": "#/components/schemas/ApiContextV1"
                        }
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ApiLaunchedV1"
                    }
                  }
                },
                "example": {
                  "request_id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                  "context": null,
                  "data": {
                    "code": "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx",
                    "expires_in_seconds": 60
                  }
                }
              }
            }
          },
          "default": {
            "description": "Typed safe error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Transflow-CSRF",
            "in": "header",
            "required": false,
            "description": "Required for browser session POSTs, except initial exchange",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiEmptyV1"
              },
              "example": {}
            }
          }
        }
      }
    },
    "/api/v1/sessions/exchange": {
      "post": {
        "operationId": "post_api_v1_sessions_exchange",
        "responses": {
          "200": {
            "description": "Context-bound result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "request_id",
                    "context",
                    "data"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "request_id": {
                      "$ref": "#/components/schemas/Uuid"
                    },
                    "context": {
                      "oneOf": [
                        {
                          "type": "null"
                        },
                        {
                          "$ref": "#/components/schemas/ApiContextV1"
                        }
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ApiSessionV1"
                    }
                  }
                },
                "example": {
                  "request_id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                  "context": null,
                  "data": {
                    "csrf": "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx",
                    "expires_in_seconds": 43200
                  }
                }
              }
            }
          },
          "default": {
            "description": "Typed safe error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Transflow-CSRF",
            "in": "header",
            "required": false,
            "description": "Required for browser session POSTs, except initial exchange",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiExchangeV1"
              },
              "example": {
                "code": "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/verify": {
      "post": {
        "operationId": "post_api_v1_sessions_verify",
        "responses": {
          "200": {
            "description": "Context-bound result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "request_id",
                    "context",
                    "data"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "request_id": {
                      "$ref": "#/components/schemas/Uuid"
                    },
                    "context": {
                      "oneOf": [
                        {
                          "type": "null"
                        },
                        {
                          "$ref": "#/components/schemas/ApiContextV1"
                        }
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ApiVerifiedV1"
                    }
                  }
                },
                "example": {
                  "request_id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                  "context": null,
                  "data": {
                    "authenticated": true
                  }
                }
              }
            }
          },
          "default": {
            "description": "Typed safe error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Transflow-CSRF",
            "in": "header",
            "required": false,
            "description": "Required for browser session POSTs, except initial exchange",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiEmptyV1"
              },
              "example": {}
            }
          }
        }
      }
    },
...
    "/api/v1/read": {
      "post": {
        "operationId": "post_api_v1_read",
        "responses": {
          "200": {
            "description": "Context-bound result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "request_id",
                    "context",
                    "data"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "request_id": {
                      "$ref": "#/components/schemas/Uuid"
                    },
                    "context": {
                      "oneOf": [
                        {
                          "type": "null"
                        },
                        {
                          "$ref": "#/components/schemas/ApiContextV1"
                        }
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ExecutionJsonV1"
                    }
                  }
                },
                "example": {
                  "request_id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                  "context": null,
                  "data": null
                }
              }
            }
          },
          "default": {
            "description": "Typed safe error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Transflow-CSRF",
            "in": "header",
            "required": false,
            "description": "Required for browser session POSTs, except initial exchange",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiReadV1"
              },
              "example": {
                "path": "example",
                "query": {}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Uuid": {
        "type": "string",
        "format": "transflow-uuid"
      },
...
      "ApiContextV1": {
        "type": "object",
        "properties": {
          "workspace": {
            "$ref": "#/components/schemas/Uuid"
          },
          "branch": {
            "type": "string",
            "maxLength": 255
          },
          "source": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Uuid"
              }
            ]
          },
          "registry": {
            "$ref": "#/components/schemas/Sha256"
          },
          "runtime_revision": {
            "type": "string",
            "maxLength": 4096
          },
          "configuration": {
            "$ref": "#/components/schemas/Sha256"
          },
          "selection": {
            "type": "object",
            "properties": {
              "kind": {
                "enum": [
                  "retained_current",
                  "plan",
                  "local_version",
                  "foreign_version"
                ]
              },
              "id": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/Uuid"
                  }
                ]
              },
              "digest": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/Sha256"
                  }
                ]
              }
            },
            "required": [
              "kind",
              "id",
              "digest"
            ],
            "additionalProperties": false
          },
          "graph": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Sha256"
              }
            ]
          },
          "freshness": {
            "const": "unknown"
          },
          "fingerprint": {
            "$ref": "#/components/schemas/Sha256"
          },
          "fallback_policy": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 1025,
            "minItems": 0
          }
        },
        "required": [
          "workspace",
          "branch",
          "source",
          "registry",
          "runtime_revision",
          "configuration",
          "selection",
          "graph",
          "freshness",
          "fingerprint",
          "fallback_policy"
        ],
        "additionalProperties": false
      },
...
      "ApiEmptyV1": {
        "type": "object",
        "properties": {},
        "required": [],
        "additionalProperties": false
      },
...
      "ApiCapabilitiesV1": {
        "type": "object",
        "properties": {
          "api_version": {
            "const": 1
          },
          "protocol_major": {
            "const": 1
          },
          "workspace": {
            "$ref": "#/components/schemas/Uuid"
          },
          "operations": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 4096
            },
            "maxItems": 200,
            "minItems": 0
          },
          "engines": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 4096
            },
            "maxItems": 200,
            "minItems": 0
          },
          "limits": {
            "type": "object",
            "properties": {
              "body_bytes": {
                "type": "integer",
                "minimum": 0,
                "maximum": 4294967295
              },
              "in_flight": {
                "type": "integer",
                "minimum": 0,
                "maximum": 4294967295
              },
              "page_default": {
                "type": "integer",
                "minimum": 0,
                "maximum": 4294967295
              },
              "page_max": {
                "type": "integer",
                "minimum": 0,
                "maximum": 4294967295
              },
              "graph_default": {
                "type": "integer",
                "minimum": 0,
                "maximum": 4294967295
              },
              "graph_max": {
                "type": "integer",
                "minimum": 0,
                "maximum": 4294967295
              },
              "graph_expansion_threshold": {
                "type": "integer",
                "minimum": 0,
                "maximum": 4294967295
              },
              "source_bytes": {
                "type": "integer",
                "minimum": 0,
                "maximum": 4294967295
              }
            },
            "required": [
              "body_bytes",
              "in_flight",
              "page_default",
              "page_max",
              "graph_default",
              "graph_max",
              "graph_expansion_threshold",
              "source_bytes"
            ],
            "additionalProperties": false
          },
          "metadata_reads_import_code": {
            "const": false
          },
          "external_reads": {
            "const": "provider_owned_leased"
          },
          "schedules": {
            "const": false
          },
          "ui": {
            "const": false
          }
        },
        "required": [
          "api_version",
          "protocol_major",
          "workspace",
          "operations",
          "engines",
          "limits",
          "metadata_reads_import_code",
          "external_reads",
          "schedules",
          "ui"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tf_session"
      }
    }
  }
}